code optimization done for resto card component and resto menu component.
