+++ src/Helper/utilityFunctions.js
@@ -1,10 +1,12 @@
 import React, { useEffect, useState } from "react";
-import { MENUAPI } from "../utils/constants";
+import { MENUAPI, SWIGGYAPIV4 } from "../utils/constants";
 import { useParams } from "react-router";
 
 // Fecthing data from SWIGGY API restaurent menu
 export const useRestaurentMenu = () => {
   const [menuData, setMenuData] = useState(null);
+  const [ListOfRatedRes, setListOfRatedRes] = useState([]);
+  const [filteredRestaurents, setFilteredRestaurents] = useState([]);
   const { resId } = useParams();
   // if I dont add the dependency array useffect will be called everytime a component renders.
   useEffect(() => {
@@ -14,9 +16,30 @@ export const useRestaurentMenu = () => {
   const fetchMenu = async () => {
     const menuApiData = await fetch(MENUAPI + resId);
     const jsonData = await menuApiData.json();
+    const data = await fetch(SWIGGYAPIV4);
+    const json = await data.json();
+    // Extracting the list of cards from the data.
+    // const cardsListV3 =
+    //   json?.data?.cards[1]?.groupedCard?.cardGroupMap?.RESTAURANT?.cards;
+    const cardListV4 =
+      json?.data?.cards[1]?.groupedCard?.cardGroupMap?.RESTAURANT?.cards[1]
+        ?.card.card?.restaurants;
+    setListOfRatedRes(cardListV4);
+    setFilteredRestaurents(cardListV4);
     setMenuData(jsonData.data);
   };
-  return menuData;
+
+  const searchHandler = () => {
+    const filteredRestaurentsList = ListOfRatedRes.filter((res) => {
+      return res.info.name
+        .toLowerCase()
+        .includes(searchInputText.toLowerCase());
+    });
+  
+    setFilteredRestaurents(filteredRestaurentsList);
+  };
+  
+  return { menuData, ListOfRatedRes, filteredRestaurents };
 };
 
 // Menu List items function
@@ -41,3 +64,6 @@ export const menuListFunction = (menuData) => {
     menuListItemsCard,
   };
 };
+
+
+
